+++ tests/testthat/test-sector_classifications.R
@@ -1,13 +1,5 @@
-test_that("hasn't change", {
-  expect_known_value(
-    sector_classifications, "ref-sector_classifications",
-    update = FALSE
-  )
-})
-
-test_that("is not different compared to reference", {
-  reference <- readRDS(test_path("ref-sector_classifications"))
-  expect_identical(sector_classifications, reference)
+test_that("hasn't changed", {
+  expect_snapshot_value(sector_classifications, style = "json2")
 })
 
 test_that("has an entry in data_dictionary", {