+++ tests/testthat/test-utils.R
@@ -25,6 +25,6 @@ test_that("define() produces the expected output", {
     "* `region` (character): Benchmark region name.",
     "* `source` (character): Source publication from which the regions are defined."
   )
-  actual <-  unclass(define("region_isos"))
+  actual <- unclass(define("region_isos"))
   expect_equal(actual, expected)
 })